[review] 팩토리 쪼개기
