+++ Drawingapp/Drawingapp/Sources/Common/Factory/TypeData/SizeFactory.swift
@@ -1,0 +1,14 @@
+//
+//  SizeFactory.swift
+//  Drawingapp
+//
+//  Created by seongha shin on 2022/03/03.
+//
+
+import Foundation
+
+class SizeFactory {
+    func make() -> Size{
+        Size(width: 150, height: 120)
+    }
+}
